Add syntax for Const different than ConstAssign
ConstAssign is a stmt, but we added syntax for a Const expr. We
basically added another level where you can look for a Const expr
so that when you run symtable_visit_expr, you can have a Const_kind
of expr.  We want to check for Const_kind in symtable.c because
symtable hands the scope of variables and could be used to prevent
reassignment of const.